Moves @NullMarked to package-info.java
See gh-3943
