+++ spring-cloud-gateway-server-webflux/src/main/java/org/springframework/cloud/gateway/filter/GlobalFilter.java
@@ -16,7 +16,6 @@
 
 package org.springframework.cloud.gateway.filter;
 
-import org.jspecify.annotations.NullMarked;
 import reactor.core.publisher.Mono;
 
 import org.springframework.web.server.ServerWebExchange;
@@ -33,7 +32,6 @@ import org.springframework.web.server.ServerWebExchange;
  * @author Rossen Stoyanchev
  * @since 5.0
  */
-@NullMarked
 public interface GlobalFilter {
 
 	/**
