+++ spring-cloud-gateway-server-webflux/src/main/java/org/springframework/cloud/gateway/filter/factory/package-info.java
@@ -14,27 +14,10 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gateway.filter;
-
-import reactor.core.publisher.Mono;
-
-import org.springframework.web.server.ServerWebExchange;
-
 /**
- * Contract to allow a {@link GatewayFilter} to delegate to the next in the chain.
- *
- * Copied from framework WebFilterChain
- *
- * @author Rossen Stoyanchev
- * @since 5.0
+ * TODO: package javadoc.
  */
-public interface GatewayFilterChain {
+@NullMarked
+package org.springframework.cloud.gateway.filter.factory;
 
-	/**
-	 * Delegate to the next {@code GatewayFilter} in the chain.
-	 * @param exchange the current server exchange
-	 * @return {@code Mono<Void>} to indicate when request handling is complete
-	 */
-	Mono<Void> filter(ServerWebExchange exchange);
-
-}
+import org.jspecify.annotations.NullMarked;
