+++ spring-cloud-gateway-server-webflux/src/main/java/org/springframework/cloud/gateway/filter/headers/HttpHeadersFilter.java
@@ -18,12 +18,9 @@ package org.springframework.cloud.gateway.filter.headers;
 
 import java.util.List;
 
-import org.jspecify.annotations.NullMarked;
-
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.server.ServerWebExchange;
 
-@NullMarked
 public interface HttpHeadersFilter {
 
 	static HttpHeaders filterRequest(List<HttpHeadersFilter> filters, ServerWebExchange exchange) {
